requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.4
django-uuslug==2.0.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-slugify==6.1.2
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
